Add support for more types
Now more types are supported in wundergraph. This commit adds support
for Uuid and NaiveDateTime, but also for self written types (like
those needed to use self written enums in postgres)
The down side of this approach is that we now generate impl's for
`LoadingHandler` and `BuildFilter` per supported backend and type and not one
generic impl per type anymore. This means this commit removes for
other backends than postgres and sqlite from `wundergraph_derive` (It
should be quite easy to add back specific backend, we just need to
generate an impl from them).
It is possible to use all supported backends together, but then only
types supported on all backends will be supported.