Remove rocket as dependency and use actix-web for the example
While this requires a bit more of code this removes the dependency to
use a certain nightly to get things working.

(This does not mean one could not use wundergraph with rocket. It only
means that it is easier for us to provide an example application that
is not tied to a specific nightly)