Use `__diesel_for_each_tuple` to implement tuple related stuff
Use a internal diesel macro to provide some impl for primary keys in the same way
diesel provides them for tuples. This also bumps the number of columns
in primary keys to whatever diesel supports maximally. (It now depents
on the "table size" feature flags on diesel. Maximally now composite
primary keys with 128 columns are supported if the right feature flag
is set)