Fix wundergraph_example to compile with the mysql backend
This requires some manual `HandleInsert` + `HandleBatchInsert` impls
because the default provided one only works with integer primary
keys, because there is no way to generally receive the last inserted
primary key value using mysql.
Interestingly we cannot derive `Insertable` then anymore because of
an error about conflicting impls with the wild card impl in
wundergraph itself. I think that's because rustc will not use
associated types to see if an impl actual is conflicting or not.